requirements.txt
# Core Framework
asyncio
asyncpg
pandas
numpy
sqlalchemy
psycopg2-binary
redis
asyncpg

# Database
psycopg2-binary
asyncpg
boto3

# Machine Learning
scikit-learn
xgboost
lightgbm
transformers
torch
sentence-transformers
tensorflow
lightgbm
fuzzywuzzy
spacy
aiofiles
asyncpg
tensorflow.keras
keras_tuner
tf-keras
seaborn

# NLP
spacy
textblob
nltk
vaderSentiment
fuzzywuzzy
python-Levenshtein

# Explainability
shap
lime

# Data processing
beautifulsoup4
lxml
feedparser
langdetect>=1.0.9

# Financial data
yfinance
alpha-vantage

# News APIs
newsapi-python

# Social media
tweepy
praw

# Web scraping
requests
selenium
scrapy
websocket-client==1.4.0

# Security
cryptography
PyJWT

# Monitoring
prometheus-client
schedule

# Configuration
PyYAML

# Testing
pytest
pytest-asyncio
pytest-cov
optuna
psutil

# Utilities
python-dotenv
pydantic
jsonschema
joblib
openai